+++ rust-src/concordium_base/src/pedersen_commitment/randomness.rs
@@ -4,6 +4,7 @@
 //! The randomness used in commitment
 
 use crate::{common::*, curve_arithmetic::*};
+use std::fmt::Debug;
 
 use rand::*;
 use std::ops::Deref;