+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,10 +1,21 @@
 ## Unreleased
 
+- Changed `Debug` implementation of secret keys and cleartext values to not print the key or value.
+- Removed `ChainParameterVersionX` types and the `MintDistributionFamily`, `GASRewardsFamily` and `AuthorizationsFamily` traits and their implementations.
+- Revised `UpdateSigner` implementations not to use references, since the one method (`sign_update_hash`) already takes a reference (`&self`).
+- Made `find_authorized_keys` public for easier re-use.
+
+## 8.0.0-alpha.3 (2025-10-08)
+
 - Add support for additional fields in `TokenModuleInitializationParameters`
 - Make `name`, `metadata` and `governanceAccount` optional in `TokenModuleInitializationParameters` and `TokenModuleState`
 - Support decoding unsigned and negative CBOR bignums to fixed-size machine integers (`i8`, ..., `i64`, `u8`, ..., `u64`)
-- Support decoding indefinite length CBOR byte and text strings
+- Introduce `Encoded<A>` a wrapper for byte encoded types using Concordium custom binary serialization format (`Serial`/`Deserial`).
+- BREAKING: Field `payload` in `UpdateInstruction` is now `Encoded<UpdatePayload>` meaning the payload stays encoded and require and additional step to decode, ensuring `UpdateInstruction` can be constructed even for future unknown `UpdatePayload` variants.
+- Added a newly defined type `WasmVersionInt` to make the Wasm Version (Smart contract version) forwards compatible by storing it there as a u8.
 - Remove `CborTokenHolder` wrapper.
+- Support decoding indefinite length CBOR byte and text strings
+- Move `BakerPoolInfo` definition to concoridum-rust-sdk
 
 ## 8.0.0 (2025-08-18)